# api/src/main/resources/api.yaml
openapi: 3.0.1
info:
  version: '1.0'
  title: Product service
servers:
  - url: '/product-service/api/v1'
tags:
  - name: Product Management
    description: Operations with Product entities
paths:
  /products:
    get:
      tags:
        - Product Management
      summary: Get all products
      description: Get all products
      operationId: getAllProducts
      parameters:
        - $ref: '#/components/parameters/list'
      responses:
        '200':
          description: All existing products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
    post:
      tags:
        - Product Management
      summary: Create Product
      description: Create Product
      operationId: createProduct
      requestBody:
        $ref: '#/components/requestBodies/CreateProductDto'
      responses:
        '200':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
  /products/{id}:
    get:
      tags:
        - Product Management
      summary: Get Product
      description: Get Product by id
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK. Product succesfully returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '404':
          description: Product not found
    put:
      tags:
        - Product Management
      summary: Update Product
      description: Update Product by id
      operationId: updateProductById
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        $ref: '#/components/requestBodies/CreateProductDto'
      responses:
        '200':
          description: Ok. Updated product succesfully returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
    delete:
      tags:
        - Product Management
      summary: Delete Product
      description: Delete Product by id
      operationId: deleteProductById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK. Product succesfully deleted
components:
  parameters:
    id:
      name: id
      in: path
      description: Product ID
      required: true
      schema:
        type: string
    list:
      name: list
      in: query
      description: List of products ID
      required: false
      schema:
        type: array
        items:
          type: string
  requestBodies:
    CreateProductDto:
      description: Product request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductDto'
  schemas:
    ProductDto:
      type: object
      properties:
        id:
          description: Unique Identifier of Product.
          type: string
          readOnly: true
        type:
          $ref: '#/components/schemas/ProductType'
        name:
          description: Name of Product.
          type: string
          readOnly: true
          maxLength: 100
        cost:
          description: Cost of Product.
          type: number
          format: double
          readOnly: true
        amount:
          description: Amount of Product.
          type: integer
          format: int32
          readOnly: true

    ProductType:
      type: string
      enum:
        - FOOD
        - CLOTHES
        - FURNITURE

    PaidProductsMessage:
      type: object
      properties:
        products:
          type: string
          additionalProperties:
            type: integer